+++ src/lte/model/lte-pdcp.h
@@ -50,7 +50,7 @@ public:
    *
    * \param rnti
    */
-  void SetRnti (uint8_t rnti);
+  void SetRnti (uint16_t rnti);
 
   /**
    *
